Add Symfony's Container::get to the list of ignored functions.
This function is often a bottleneck in very fast Symfony code,
when aggregated over all calls. However slow is only the actual
instantiation of the objects. Given how XHProf Hierachical profiling
only captures one level of call depth, not ignoring this function makes
it difficult to spot which service instantiation are actually slowing
which part of the code down.

This optimization should be done for any kind of DependencyInjection
container that has the same problem, however currently its only
easy for the Symfony one, because it will call a method directly
on a subclass.